implemented code generation and queue management
