+++ oasp4j/jumpTheQueue/core/.settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<faceted-project>
+  <installed facet="jst.utility" version="1.0"/>
+  <installed facet="java" version="1.8"/>
+</faceted-project>
