--- oasp4j/jumpTheQueue/core/src/main/java/com/cap/jumpthequeue/accesscode/common/api/AccessCode.java
@@ -1,21 +1,0 @@
-package com.cap.jumpthequeue.accesscode.common.api;
-
-import java.util.Date;
-
-import com.cap.jumpthequeue.general.common.api.ApplicationEntity;
-
-public interface AccessCode extends ApplicationEntity {
-
-  public String getCode();
-
-  public void setCode(String code);
-
-  public Date getDateAndTime();
-
-  public void setDateAndTime(Date dateAndTime);
-
-  public Long getVisitorId();
-
-  public void setVisitorId(Long visitorId);
-
-}
