+++ oasp4j/jumpTheQueue/core/src/main/resources/db/migration/V0002__R001_Master_data.sql
@@ -3,5 +3,5 @@ INSERT INTO Visitor (id, modificationCounter, name, email, phone, idCode) VALUES
 INSERT INTO Visitor (id, modificationCounter, name, email, phone, idCode) VALUES (2, 1, 'Peter', 'peter@mail.com', '789101', 2);
 
 
-INSERT INTO AccessCode (id, modificationCounter, code, dateAndTime, idVisitor) VALUES (1, 1, 'A01','2017-08-01', 1);
-INSERT INTO AccessCode (id, modificationCounter, code, dateAndTime, idVisitor) VALUES (2, 1, 'A02','2017-08-02', 2);
+INSERT INTO AccessCode (id, modificationCounter, code, dateAndTime, idVisitor) VALUES (1, 1, 'A0001', CURRENT_TIMESTAMP + (60 * 60 * 24 * 5), 1);
+INSERT INTO AccessCode (id, modificationCounter, code, dateAndTime, idVisitor) VALUES (2, 1, 'A0002', CURRENT_TIMESTAMP + (60 * 60 * 24 * 5), 2);
