+++ oasp4j/jumpTheQueue/core/src/main/resources/db/migration/h2/V0001__R001_Create_schema.sql
@@ -11,6 +11,7 @@ CREATE TABLE Visitor(
   id BIGINT NOT NULL AUTO_INCREMENT,
   modificationCounter INTEGER NOT NULL,
   name VARCHAR(255),
+  surname VARCHAR(255),
   email VARCHAR(255),
   phone VARCHAR(255),
   idCode BIGINT,
@@ -23,7 +24,7 @@ CREATE TABLE ACCESSCODE(
   id BIGINT NOT NULL AUTO_INCREMENT,
   modificationCounter INTEGER NOT NULL,
   code VARCHAR(5),
-  dateAndTime timestamp,
+  dateAndTime TIMESTAMP,
   idVisitor BIGINT,
   CONSTRAINT PK_AccessCode PRIMARY KEY(id),
   CONSTRAINT FK_AccessCode_idVisitor FOREIGN KEY(idVisitor) REFERENCES Visitor(id)
