+++ oasp4j/jumpTheQueue/server/.settings/org.eclipse.wst.common.component
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
+    <wb-module deploy-name="jumpthequeue-server-0.1">
+        <wb-resource deploy-path="/" source-path="/target/m2e-wtp/web-resources"/>
+        <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
+        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
+        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
+        <dependent-module archiveName="jumpthequeue-core-0.1.jar" deploy-path="/WEB-INF/lib" handle="module:/resource/jumpthequeue-core/jumpthequeue-core">
+            <dependency-type>uses</dependency-type>
+        </dependent-module>
+        <property name="context-root" value="jumpthequeue-server"/>
+        <property name="java-output-path" value="/jumpthequeue-server/target/classes"/>
+    </wb-module>
+</project-modules>
