+++ oasp4j/jumpTheQueue/server/.settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<faceted-project>
+  <fixed facet="wst.jsdt.web"/>
+  <installed facet="java" version="1.7"/>
+  <installed facet="jst.web" version="3.0"/>
+  <installed facet="jst.jaxrs" version="2.0"/>
+  <installed facet="wst.jsdt.web" version="1.0"/>
+</faceted-project>
